StandAloneService/requirements.txt
asgiref==3.2.10
cffi==1.14.3
cryptography==3.1
Django==3.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
